Mise en place des Traits
