+++ src/Entity/ClassRoom.php
@@ -2,6 +2,8 @@
 
 namespace App\Entity;
 
+use App\Entity\Traits\IdentifierTrait;
+use App\Entity\Traits\TimestampableTrait;
 use Doctrine\ORM\Mapping as ORM;
 
 /**
@@ -9,17 +11,8 @@ use Doctrine\ORM\Mapping as ORM;
  */
 class ClassRoom
 {
-    /**
-     * @ORM\Id()
-     * @ORM\GeneratedValue()
-     * @ORM\Column(type="integer")
-     */
-    private $id;
-
-    /**
-     * @ORM\Column(type="string", length=255)
-     */
-    private $uuid;
+    use TimestampableTrait;
+    use IdentifierTrait;
 
     /**
      * @ORM\Column(type="string", length=255)
@@ -31,32 +24,6 @@ class ClassRoom
      */
     private $name;
 
-    /**
-     * @ORM\Column(type="datetime_immutable")
-     */
-    private $createdDate;
-
-    /**
-     * @ORM\Column(type="datetime")
-     */
-    private $updatedDate;
-
-    public function getId(): ?int
-    {
-        return $this->id;
-    }
-
-    public function getUuid(): ?string
-    {
-        return $this->uuid;
-    }
-
-    public function setUuid(string $uuid): self
-    {
-        $this->uuid = $uuid;
-
-        return $this;
-    }
 
     public function getGrade(): ?string
     {
@@ -78,30 +45,6 @@ class ClassRoom
     public function setName(string $name): self
     {
         $this->name = $name;
-
-        return $this;
-    }
-
-    public function getCreatedDate(): ?\DateTimeImmutable
-    {
-        return $this->createdDate;
-    }
-
-    public function setCreatedDate(\DateTimeImmutable $createdDate): self
-    {
-        $this->createdDate = $createdDate;
-
-        return $this;
-    }
-
-    public function getUpdatedDate(): ?\DateTimeInterface
-    {
-        return $this->updatedDate;
-    }
-
-    public function setUpdatedDate(\DateTimeInterface $updatedDate): self
-    {
-        $this->updatedDate = $updatedDate;
 
         return $this;
     }
